Move tzpath setting into common file
There is no particular need for this code to be duplicated in C and
Python, so we'll call into Python from C to get access to the time zone
path.
