+++ src/zoneinfo/__init__.py
@@ -1,4 +1,5 @@
 __all__ = ["ZoneInfo", "set_tzpath"]
 
+from ._tzpath import set_tzpath
 from ._version import __version__
-from ._zoneinfo import ZoneInfo, set_tzpath
+from ._zoneinfo import ZoneInfo
